review: reviewed by Fable 5

+++ src/shuiyuan_auto_reply/tarot/tarot_group_data.py
@@ -105,7 +105,7 @@ class BaseTarotGroup:
         return {}
 
     @classmethod
-    def match_score(cls, question: str) -> int:
+    def match_score(cls, question: str) -> float:
         """
         Calculate the match score for the question based on keywords.
